# 8307446: RISC-V: Improve performance of floating point to integer conversion