--- a/src/hotspot/cpu/riscv/macroAssembler_riscv.cpp
+++ b/src/hotspot/cpu/riscv/macroAssembler_riscv.cpp
@@ -4061,10 +4061,10 @@ void MacroAssembler::zero_dcache_blocks(Register base, Register cnt, Register tm
   bge(cnt, tmp1, loop);
 }
 
-#define FCVT_SAFE(FLOATCVT, IS_DOUBLE)                                                    \
+#define FCVT_SAFE(FLOATCVT, FLOATSIG)                                                     \
 void MacroAssembler::FLOATCVT##_safe(Register dst, FloatRegister src, Register tmp) {     \
   Label do_convert, done;                                                                 \
-  IS_DOUBLE ? fclass_d(tmp, src) : fclass_s(tmp, src);                                    \
+  fclass_##FLOATSIG(tmp, src);                                                            \
   /* check if src is NaN */                                                               \
   andi(tmp, tmp, 0b1100000000);                                                           \
   beqz(tmp, do_convert);                                                                  \
@@ -4075,10 +4075,10 @@ void MacroAssembler::FLOATCVT##_safe(Register dst, FloatRegister src, Register t
   bind(done);                                                                             \
 }
 
-FCVT_SAFE(fcvt_w_s, false);
-FCVT_SAFE(fcvt_l_s, false);
-FCVT_SAFE(fcvt_w_d, true);
-FCVT_SAFE(fcvt_l_d, true);
+FCVT_SAFE(fcvt_w_s, s);
+FCVT_SAFE(fcvt_l_s, s);
+FCVT_SAFE(fcvt_w_d, d);
+FCVT_SAFE(fcvt_l_d, d);
 
 #undef FCVT_SAFE