@lgxbslgx 2023-05-06
@feilongjiang The command `/integrate` is [allowed in the comment body](https://github.com/openjdk/skara/blob/0b67e8fdcb1f8231b18491f8d2581573fa63c792/bots/pr/src/main/java/org/openjdk/skara/bots/pr/IntegrateCommand.java#L379) if it is [started at a new line](https://github.com/openjdk/skara/blob/0b67e8fdcb1f8231b18491f8d2581573fa63c792/bots/pr/src/main/java/org/openjdk/skara/bots/pr/CommandExtractor.java#L115). So feel free to do that.